refactor: move useBootstrapApp to react-hooks package
