+++ platforms/web/package.json
@@ -33,7 +33,6 @@
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-i18next": "^12.3.1",
-    "react-query": "^3.39.0",
     "react-router": "6.14.2",
     "react-router-dom": "6.14.2",
     "wicg-inert": "^3.1.1",
